chore(bus): refactoring

+++ tchalanet-server/tchalanet-core/src/main/java/com/tchalanet/server/core/limitpolicy/internal/application/query/handler/evaluation/EvaluateLimitPolicyQueryHandler.java
@@ -45,9 +45,9 @@ public class EvaluateLimitPolicyQueryHandler
                     scopeLabel(b.appliedScope()),
                     b.code(),
                     b.messageKey(),
-                    BigDecimal.valueOf(b.limitValue()),
-                    BigDecimal.valueOf(b.currentValue()),
-                    BigDecimal.valueOf(b.deltaValue())
+                    b.limitValue(),
+                    b.currentValue(),
+                    b.deltaValue()
                 ))
                 .toList());
     }
